+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - adding `sort` to sort a list
 - added `\t`, `\n`, `\v` and `\r` escape codes (available in strings only)
 - adding `listFiles` which returns a list of paths (strings)
+- adding `(makeDir path)` and `(removeFiles ...)`
 
 ### Changed
 - updated output of tests with number of passed tests